+++ src/main/java/duke/Duke.java
@@ -1,8 +1,7 @@
+package duke;
+
 import java.io.IOException;
 import java.util.Scanner;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeFormatterBuilder;
-import java.time.temporal.ChronoField;
 
 public class Duke {
