+++ src/main/java/duke/Storage.java
@@ -1,11 +1,11 @@
+package duke;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Scanner;
-import java.io.FileNotFoundException;
-
 
 
 public class Storage {
