+++ src/main/java/duke/Todo.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class Todo extends Task {
     Todo(String desc) {
         super(desc);
